encapsulate bip44 logic into a new class BIP44
